fix: arruma erros de typescript
