+++ src/App.css
@@ -100,7 +100,7 @@ section{
   }
 }
  
-@layer base {
+/* @layer base {
   * {
     @apply border-border;
   }
@@ -108,4 +108,4 @@ section{
     @apply bg-background text-foreground;
     font-feature-settings: "rlig" 1, "calt" 1;
   }
-}
+} */
