+++ src/components/basicosComponents/CardAdminCardapio/index.jsx
@@ -20,7 +20,7 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-} from "src/componentes/card";
+} from "../../../componentes/card";
 
 /* const est_id = localStorage.getItem("est_id"); */
 const url = urlApi;
