--- src/components/basicosComponents/DivInput/index.jsx
@@ -1,8 +1,0 @@
-export function DivInput({children, ...props}) {
-    const className = props.className
-    return (
-        <div className={`flex flex-col justify-center w-full gap-1 ${className}`}>
-            {children}
-        </div>
-    )
-}
