--- src/components/basicosComponents/Label/index.jsx
@@ -1,3 +1,0 @@
-export function Label({ children, htmlFor }) {
-  return <label htmlFor={htmlFor} className="font-medium">{children}</label>;
-}
