+++ src/components/basicosComponents/Label/index.tsx
@@ -1,0 +1,10 @@
+import React, { ReactNode } from 'react';
+
+interface LabelProps {
+    children?: ReactNode;
+    htmlFor: string;
+}
+
+export function Label({ children, htmlFor }: LabelProps) {
+    return <label htmlFor={htmlFor} className="font-medium">{children}</label>;
+}
